Parsing UTF-8 responses
By using the `json()` method on the Response object, we can get the unpacked
representation of the JSON response properly decoded into unicode strings.

Edited a response from the stub server to ensure that Crowd doesn't fail
with the response, as well as an explicit test to ensure the returned
string actually contains the UTF-8 characters.